refactor(C): v0.2.7 C7f — _format_steady_block/_format_ev_ebitda_block 迁入 valuation_calc.py
- 两函数逐字节原样搬迁（保留 dict 签名与 docstring），同源渲染器不再拆两文件
- invest.py cmd_value 的 import 追加两名，三段 print 顺序不变
- 验证：临时基线 10 变体（format 3 + steady 4 + ev 3）搬迁前后逐字节 0 diff；
  test_valuation + test_invest_cli_help 全过；全仓 grep 无其他调用者
